added new tools to Kipon.Xrm.Cmd : exportassmsteps and importassmsteps. Documentation will follow in next normal release of the plugin framework.
The tool allow you to export all step and image definition from one environemnt (ex. a managed production environment) into a developer instance unmanaged This is convinient if you by mistake unregistre all steps for a developer instance, and dont have a propert backup, so the source of true is your prod environment.